requirements.txt
python-dotenv
openai
langchain
langchain-openai
langchain-community
langgraph
tavily-python
duckduckgo-search
beautifulsoup4
Pygments